Add endpoint for fetching specific resident by id
Look up the resident with a find. Not the most performant, but since
the data set is fixed I'm not worried about it. An easy fix would be
to organize the data into a map when the server boots so that we can
do an O(1) lookup.

Handling errors in Deno seems nice. Just throwing an instance of the
relevant httpError with a message.