Add endpoint for recommendations
Can visit /api/resident/:id/recommended-programs to get three programs
recommended for that resident.

I wanted to do recommendations around the events based on their names,
but I concluded that, for that to be good, it would require being able
to recognize two names that were similar.

Then I though maybe hobbies would be a good focus, but those also seemed
to be unreliable especially since some resident's hobbies to not have
any related programs that have run in the past.

From there, dimension seemed to be the most reliable attribute on an
event, so I choose to focus on that. Level of care for a resident also
seemed very important since if the event does not accomodate their level
of care it probably wouldn't work well for them.

Based on a resident's attendance history, I find the most popular
dimension of program they attend. I then pick a program they have
attended that meets that dimension and that they have attended in the
past. Then I do the same based on their second most popular dimension.
Finally, I pick a third based on their most popular dimension but that
they have not attended to give more variety.

Things to improve:

  - Could randomize the results to give some variety
  - Strip some data off of the returned programs that are specific
    to the instance of it (i.e. attendees, start, end, id)
  - Would be awesome to create new events that made sense. Would likely
    require some level of AI/ML.
  - Could look at other ways to recommend based on activity of other
    residents that have similar attendance habits
    - I had initially investigated this direction, but I was trying to
    categorize the programs to group similar ones together and that is
    when I noticed all the tiny variations in the data of similar
    programs. I would need to ignore that grouping, create a history
    matrix, then a co-occurance matrix and use that to come up with
    recommendations that would be new to the resident. Admittedly, this
    would be better if we were showing these recommendations to the
    residents for helping them choose what to attend, not the staff for
    helping them to choose which programs to put on.